refactor: create api blueprint for modal/forms
